cleaned saga
