+++ src/components/LoginPage/LoginPage.js
@@ -1,14 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-
 import Teams from '../Teams/Teams';
 import UserProfile from '../UserProfile/UserProfile';
 
-//I (Haley) changed this file a lot from the prime starter
-//for testing purposes and easier prs I wanted everything in one page
-//I think once things are working it would make a lot of sense to make it a lot more modular
-//for example prime code has a login page and a register page
-//i'm just conditionally rendering and doing functions based on this.state.mode
 
 class LoginPage extends Component {
   state = {
@@ -27,14 +21,13 @@ class LoginPage extends Component {
     this.props.dispatch({ type: 'FETCH_CURRENT_USER' });
     
 
-  }//this gets user info, upon reload (from session, I think)
+  }//this gets user info, upon reload
 
   login = (event) => {
     //don't reload page on submit
     event.preventDefault();
     //if all the things you need are filled for the current mode (this may change) dispatch the state to a saga
     //otherwise, let the user know what they did wrong 
-    //(prime code has errors from reducers, we didn't copy that over so I'm gonna use a general alert for now)
     if (this.state.mode==='login' && this.state.email && this.state.password) {
       this.props.dispatch({
         type: 'LOGIN',
@@ -44,15 +37,7 @@ class LoginPage extends Component {
         },
       });
       this.props.history.push('/home');
-      // this.props.history.push({ pathname: '/' })
-      // if (this.props.user.currentUser.level === 1) {
-      //   this.props.history.push({ pathname: '/' })
-      // }
-      // if (this.props.user.currentUser.level === 2) {
-      //   this.props.history.push({ pathname: '/clientHome' })
-      // }
     }//end if login mode
-    //register will eventually have more required fields but for now just email and password
     else if (this.state.mode === 'register' && this.state.email && this.state.password) {
       this.props.dispatch({
         type: 'REGISTER',
@@ -67,12 +52,6 @@ class LoginPage extends Component {
         },
       });
       this.props.history.push({ pathname: '/' })
-      // if(this.props.user.currentUser.level === 1){
-      //   this.props.history.push({ pathname: '/' })
-      // }
-      //   if(this.props.user.currentUser.level === 2){
-      //     this.props.history.push({ pathname: '/clientHome' })
-      // }
     }//end if register mode
     else if (this.state.mode ==='login'){
       this.props.dispatch({ type: 'LOGIN_INPUT_ERROR' });
@@ -81,32 +60,26 @@ class LoginPage extends Component {
       this.props.dispatch({ type: 'REGISTRATION_INPUT_ERROR' });
     }//end else (fields not filled, not able to dispatch)
     // window.location.reload(false);
-
   }//end login
 
   handleInputChangeFor = propertyName => (event) => {
     this.setState({
       [propertyName]: event.target.value,
     });
-  }//store input- didn't change from prime starter
+  }
 
+  // switches between login mode and register mode
   swtichMode=()=>{
-    //update the state so new things show for buttons, header, and input
-    //the copy is a clumsy/not for production, should be changed
-    //also I think buttons and h1 text should maybe be conditionally rendered instead of from state
     if (this.state.mode==='login'){
       this.setState({mode:'register', toggleButtonText:'switch to login mode', headerText: 'Sign Up', submitValue: 'Register'})}
     else{
       this.setState({ mode: 'login', toggleButtonText: 'switch to register mode', headerText: 'Log In', submitValue: 'Log In' })}
   }
 
-  //rendering login stuff vs register stuff depending on local state
-  //probably should be broken into login page vs register page eventually
-  //at least it'll be easy to copy the code with it divided like this?
+ 
   render() {
     return (
       <div>
-        {/* <p>this.props.user.currentUser: {JSON.stringify(this.props.user)}</p> */}
         {this.props.user.currentUser.id 
           ?
             <div>
@@ -116,7 +89,7 @@ class LoginPage extends Component {
             </div>
           :
             <>
-            {/* display this stuff when there isn't a user detected */}
+            {/* display this when there isn't a user detected */}
               {this.state.mode==='login'
                 ?
                 <>
@@ -209,5 +182,3 @@ const mapStateToProps = state => ({
 });
 
 export default connect(mapStateToProps)(LoginPage);
-//prime code maps errors reducer to props so that dom displays different things accordingly
-//seems like a good idea for user experience but not core to 'getting auth to work'
