+++ src/components/WorkflowEditor/AddTask/subcomponents/AddTask2.js
@@ -12,12 +12,12 @@ class AddTask2 extends Component {
     nextStep = () => {
         this.props.dispatch({ type: 'SET_TASK_DESCRIPTION', payload: this.state.descriptionInput });
         this.props.dispatch({ type: 'NEXT_TASK_STEP' });
-        // this.props.history.push('/add-task-haley/3')
+
     }
 
     goBack = () => {
         this.props.dispatch({ type: 'PREVIOUS_TASK_STEP' });
-        // this.props.history.push('/add-task-haley/1')
+
     }
 
     handleDescriptionInput =(event)=> {
