+++ src/redux/sagas/workflowSaga.js
@@ -20,6 +20,18 @@ function* workflows() {
     yield takeEvery('ADD_NEW_TASK', addNewTask);
     yield takeEvery('GET_TASK_OPTIONS', getTaskOptions);
     yield takeEvery('REMOVE_TASK', removeTask);
+    // go through below for duplication
+    yield takeEvery('SET_TASK_PHASE_ID', setTaskPhaseID);
+    yield takeEvery('FETCH_TASK_RISK_TYPES', setTaskRiskareaOptions);
+    yield takeEvery('SET_TASK_TITLE', setTaskTitle);
+    yield takeEvery('SET_TASK_RISKAREAS', setTaskRiskareas);
+    yield takeEvery('SET_TASK_DESCRIPTION', setTaskDescription);
+    yield takeEvery('SET_TASK_SEQUENCE', setTaskSequence);
+    yield takeEvery('ADD_TASK_LINK', addTaskLink);
+    yield takeEvery('ADD_TASK_INPUT', addTaskInput);
+    yield takeEvery('ADD_TASK_TO_DATABASE', addTaskToDatabase);
+    yield takeEvery('ADD_TASK_TO_SUBSCRIBER_DATABASE', addTaskToSubscriberDatabase);
+    yield takeEvery('FETCH_ASSIGNED_TASK', fetchAssignedTask);
 }
 
 // gets all workflows from DB
@@ -194,5 +206,82 @@ function* getTaskOptions(){
     console.log('in saga - all task options GET back with:', getOptions.data);
     yield put({type: 'SET_TASK_OPTIONS', payload: getOptions.data});
 }
+
+function* setTaskPhaseID (action) {
+    yield put({type: 'UPDATE_PHASE_ID', payload: action.payload});
+}
+
+function* setTaskRiskareaOptions(action) {
+    const riskTypes = yield axios.get(`/api/workflow/riskareas/${action.payload}`);
+    let riskTypesArray = [];
+    for (let i = 0; i < riskTypes.data.length; i++) {
+        // console.log(riskTypes.data[i].riskarea)
+        riskTypesArray.push(riskTypes.data[i]);
+    }
+    yield put({ type: 'UPDATE_RISKAREA_OPTIONS', payload: riskTypesArray });
+}
+
+function* setTaskTitle(action) {
+    yield put({ type: 'UPDATE_TASK_TITLE', payload: action.payload });
+}
+
+function* setTaskRiskareas(action) {
+    yield put({ type: 'UPDATE_TASK_RISKAREAS', payload: action.payload });
+}
+
+function* setTaskDescription(action) {
+    yield put({ type: 'UPDATE_TASK_DESCRIPTION', payload: action.payload });
+}
+
+function* setTaskSequence(action){
+    yield put({ type: 'UPDATE_TASK_SEQUENCE', payload: action.payload})
+}
+
+function* addTaskLink(action) {
+    yield put({ type: 'UPDATE_TASK_LINKS', payload: action.payload });
+} 
+
+function* addTaskInput(action) {
+    yield put({ type: 'UPDATE_TASK_INPUTS', payload: action.payload });
+}
+
+function* addTaskToDatabase(action) {
+    try {
+        const postResponse = yield axios.post(`/api/workflow/add-new-task`, action.payload.task);
+        console.log('postResponse:', postResponse);
+        yield put({ type: 'ADD_TASK_CONFIRM', payload: postResponse.data.newTaskId });
+        yield put({type: 'GET_THIS_PHASE', payload: action.payload});
+        yield put({type: 'SET_TASK_PHASE_ID', payload: action.payload.id});
+    } catch (error) {
+        console.log('error with posting new task:', error);
+    }
+}
+
+function* addTaskToSubscriberDatabase(action) {
+    try {
+        const postResponse = yield axios.post(`/api/workflow/add-new-assigned-task`, action.payload);
+        console.log('postResponse:', postResponse);
+        yield put({ type: 'ADD_TASK_CONFIRM', payload: postResponse.data.newTaskId });
+    } catch (error) {
+        console.log('error with posting new task:', error);
+    }
+}
+
+function* fetchAssignedTask(action) {
+    try {
+        // let assignedTaskObject = {}
+        const firstResponse = yield axios.get(`/api/workflow/assigned-task/${action.payload.id}`);
+        console.log('response:', firstResponse);
+        // assignedTaskObject= firstResponse[0];
+        const secondResponse = yield axios.get(`/api/workflow/assigned-task/inputs/${action.payload.id}`);
+        console.log('second response:', secondResponse);
+        //get all the things from the other tables and bundle into one task object that subscriber dom can render from 
+        // yield put({ type: 'SET_ASSIGNED_TASK', payload: response.data });
+    } catch (error) {
+        console.log('error with posting new task:', error);
+    }
+}
+
+
 
 export default workflows;
